requirements/base.txt
PyQt6
darkdetect
attrs
cattrs
dataset
